# Conflicts:
#	yarn.lock

+++ src/app/features/tasks/dialog-add-task-reminder/dialog-add-task-reminder.component.ts
@@ -1,5 +1,5 @@
 import { ChangeDetectionStrategy, Component, Inject } from '@angular/core';
-import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
+import { MAT_DIALOG_DATA, MatDialog, MatDialogRef } from '@angular/material/dialog';
 import { TaskService } from '../task.service';
 import { ReminderCopy } from '../../reminder/reminder.model';
 import { ReminderService } from '../../reminder/reminder.service';
@@ -9,6 +9,9 @@ import { throttle } from 'helpful-decorators';
 import { Task, TaskReminderOption, TaskReminderOptionId } from '../task.model';
 import { millisecondsDiffToRemindOption } from '../util/remind-option-to-milliseconds';
 import { LS_LAST_IS_MOVE_SCHEDULED_TO_BACKLOG } from '../../../core/persistence/ls-keys.const';
+import { isToday } from '../../../util/is-today.util';
+import { DialogConfirmComponent } from '../../../ui/dialog-confirm/dialog-confirm.component';
+import { TASK_REMINDER_OPTIONS } from './task-reminder-options.const';
 
 @Component({
   selector: 'dialog-add-task-reminder',
@@ -29,40 +32,13 @@ export class DialogAddTaskReminderComponent {
     !!this.task.projectId && this.task.parentId === null && !this.task.repeatCfgId;
   isMoveToBacklog: boolean;
   // TODO make translatable
-  remindAvailableOptions: TaskReminderOption[] = [
-    {
-      // id: TaskReminderOptionId.DoNotRemind,
-      // title: 'Dont show reminder',
-      // }, {
-      id: TaskReminderOptionId.AtStart,
-      title: T.F.TASK.D_REMINDER_ADD.RO_START,
-    },
-    {
-      id: TaskReminderOptionId.m5,
-      title: T.F.TASK.D_REMINDER_ADD.RO_5M,
-    },
-    {
-      id: TaskReminderOptionId.m10,
-      title: T.F.TASK.D_REMINDER_ADD.RO_10M,
-    },
-    {
-      id: TaskReminderOptionId.m15,
-      title: T.F.TASK.D_REMINDER_ADD.RO_15M,
-    },
-    {
-      id: TaskReminderOptionId.m30,
-      title: T.F.TASK.D_REMINDER_ADD.RO_30M,
-    },
-    {
-      id: TaskReminderOptionId.h1,
-      title: T.F.TASK.D_REMINDER_ADD.RO_1H,
-    },
-  ];
+  remindAvailableOptions: TaskReminderOption[] = TASK_REMINDER_OPTIONS;
   selectedReminderCfgId: TaskReminderOptionId;
 
   constructor(
     private _taskService: TaskService,
     private _reminderService: ReminderService,
+    private _matDialog: MatDialog,
     private _matDialogRef: MatDialogRef<DialogAddTaskReminderComponent>,
     @Inject(MAT_DIALOG_DATA) public data: AddTaskReminderInterface,
   ) {
@@ -112,17 +88,44 @@ export class DialogAddTaskReminderComponent {
       });
       this.close();
     } else {
-      this._taskService.scheduleTask(
-        this.task,
-        timestamp,
-        this.selectedReminderCfgId,
-        this.isMoveToBacklog,
-      );
-      localStorage.setItem(
-        LS_LAST_IS_MOVE_SCHEDULED_TO_BACKLOG,
-        this.isMoveToBacklog + '',
-      );
-      this.close();
+      if (!!this.task.repeatCfgId && !isToday(timestamp)) {
+        this._matDialog
+          .open(DialogConfirmComponent, {
+            restoreFocus: true,
+            data: {
+              message: T.F.TASK.D_REMINDER_ADD.CONFIRM_REPEAT_TXT,
+              okTxt: T.F.TASK.D_REMINDER_ADD.CONFIRM_REPEAT_OK,
+            },
+          })
+          .afterClosed()
+          .subscribe((isConfirm: boolean) => {
+            if (isConfirm) {
+              this._taskService.scheduleTask(
+                this.task,
+                timestamp,
+                this.selectedReminderCfgId,
+                this.isMoveToBacklog,
+              );
+              localStorage.setItem(
+                LS_LAST_IS_MOVE_SCHEDULED_TO_BACKLOG,
+                this.isMoveToBacklog + '',
+              );
+              this.close();
+            }
+          });
+      } else {
+        this._taskService.scheduleTask(
+          this.task,
+          timestamp,
+          this.selectedReminderCfgId,
+          this.isMoveToBacklog,
+        );
+        localStorage.setItem(
+          LS_LAST_IS_MOVE_SCHEDULED_TO_BACKLOG,
+          this.isMoveToBacklog + '',
+        );
+        this.close();
+      }
     }
   }
